pyproject.toml
[project]
name = "hangoutsscheduler"
version = "0.1.0"
description = ""
authors = [
    {name = "Ryan Sandoval",email = "ryan_sandoval@live.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=0.3.15,<0.4.0)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "langgraph (>=0.2.70,<0.3.0)",
    "discord-py (>=2.3.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
